# services/gateway/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gazelle Public API Gateway",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/orders/{orderId}/pay": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/orders/{orderId}/cancel": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.gazellecoffee.com/v1"
    }
  ]
}